requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
aiohttp==3.8.1
pydantic==1.8.2
python-dotenv==0.19.0
redis==4.2.2
ratelimit==2.2.1